requirements.txt
dash
transformers
torch
python-dotenv
gunicorn